+++ src/ScoreCalculator/Branch.java
@@ -1,20 +1,19 @@
 package src.ScoreCalculator;
 
-
 class Branch {
     int[] realTaxaCountsTotal;
-    int[] dummyTaxaCountsTotal;
-    int[] dummyTaxaCountsIndividual;
+    double[] dummyTaxaCountsTotal;
+    double[] dummyTaxaCountsIndividual;
     int[] dummyTaxaToPartitionMap;
-    int[] totalTaxaCounts;
+    double[] totalTaxaCounts;
 
 
-    Branch(int[] rtc, int[] dtci, int[] dtct, int[] dtpm) {
+    Branch(int[] rtc, double[] dtci, double[] dtct, int[] dtpm) {
         this.realTaxaCountsTotal = rtc;
         this.dummyTaxaCountsIndividual = dtci;
         this.dummyTaxaCountsTotal = dtct;
         this.dummyTaxaToPartitionMap = dtpm;
-        this.totalTaxaCounts = new int[2];
+        this.totalTaxaCounts = new double[2];
         for(int i = 0; i < 2; ++i){
             this.totalTaxaCounts[i] = this.realTaxaCountsTotal[i] + this.dummyTaxaCountsTotal[i];
         }
